parser: relocate geometry decode (layout, wire_table) from render/ into parser/
The renderer was reaching past the parser to read heap XML directly
(render/layout.py did its own ET.parse + extract_vi_xml). That inverts the
intended layering: the parser is lvkit's single XML abstraction; render should
consume geometry, not read it.

Pure move — layout.py and wire_table.py relocate render/ -> parser/ with only
import-path updates (render/scene.py, render/glyph.py, tests, one script). No
logic change. Verified render output byte-identical across a 45-heap spread and
the full render suite green (136 passed). This is the mechanical first step;
parser producing the Layout in its own single read follows.
