+++ scripts/wire_decode_probe.py
@@ -33,8 +33,8 @@ import xml.etree.ElementTree as ET
 from pathlib import Path
 
 from lvkit.extractor import extract_vi_xml
-from lvkit.render.layout import _LayoutBuilder
-from lvkit.render.wire_table import DIR, _decode_chain, _decode_lengths, _flip
+from lvkit.parser.layout import _LayoutBuilder
+from lvkit.parser.wire_table import DIR, _decode_chain, _decode_lengths, _flip
 
 Point = tuple[float, float]
 N, S, E, W = (0, -1), (0, 1), (1, 0), (-1, 0)
